% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_functions.R
\name{addCovariance}
\alias{addCovariance}
\title{Computes covariance matrices from available variogram models within the fstat structure}
\usage{
addCovariance(.g = NULL, type = "omni")
}
\arguments{
\item{.g}{-  An 'fstat' structure}

\item{type}{-  The type of covariance to compute ('omni', 'separable', 'sum', 'product', 'productSum')}
}
\details{
It will compute a covariance matrix for each variable in the fstat structure.
         Note: 'separable, 'product', 'sum', productSum' are experimental in the current version.
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}
