% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_functions.R
\name{estimateDrift}
\alias{estimateDrift}
\title{The function estimates functional or scalar drift.}
\usage{
estimateDrift(.formula, .g, .type = "OLS", Intercept = TRUE, regCoef = 0)
}
\arguments{
\item{.formula}{- formula specifying which variables to use for drift;}

\item{.g}{- an fstat structure containing all data.}

\item{.type}{- drift type (OLS, GLS, MEAN)}

\item{.regCoef}{- Regularization coefficient (just like in ridge regression)}
}
\details{
It adds a "drift" slot to the fstat structure (.g) with the following components:
         Betas         - Coefficients assigned to each variable specified by the formula;
         Design matrix - A design matrix for each variable (useful for UK);
         Residuals     - Functional or scalar residuals;

         This code does not implement functional regression (Ramsay and Silverman, 2005), instead it fits one regression
         for each time step (piece-wise). The actual functional regression will be implemented in future version(s).

         for .type=MEAN it just subtracts the mean from the functional data.
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu}) and Alessandra Menafoglio (\email{alessandra.menafoglio@polimi.it})
}
