% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_functions.R
\name{finterpfitLMC}
\alias{finterpfitLMC}
\title{A modified version of the 'fit.lmc' function from the gstat package}
\usage{
finterpfitLMC(.v, .n, .model, fit.ranges = FALSE, fit.lmc = !fit.ranges,
  correct.diagonal = 1, fNugget = FALSE, ...)
}
\details{
Not to be called directly. This is an internal function!
}
\author{
Modified by: Ogy Grujic (\email{ogyg@stanford.edu})
}
