% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_functions.R
\name{fitVariograms}
\alias{fitVariograms}
\title{A function that fits variograms to empirical estimates}
\usage{
fitVariograms(g = NULL, model = NULL, dir = NULL, correctDiagonal = 1,
  fitSills = TRUE, fitRanges = TRUE, forceNugget = FALSE,
  posNugget = FALSE)
}
\arguments{
\item{g}{- An \code{fstat} structure.}

\item{model}{- Variogram model of class \code{'vgm'}.}

\item{dir}{- Direction you wish to fit. Useful for fine tunning.}

\item{correctDiagonal}{- Default = 1. Corrects the diagonal of the LMC sill matrix.}

\item{fitSills}{- This is a convenience access parameter for fit.variogram}

\item{fitRanges}{- This is a convenience access parameter for fit.variogram}

\item{forceNugget}{- Whether to force the nugget to user specified value or to fit it (Boolean).}

\item{posNugget}{- Whether to enforce the positivity of the nugget (Boolean). Not to be used
with conventional cross variogram computation!}
}
\details{
It returns the \code{fstat} structure with an addtional slot \code{models}.
         This function will cycle through all available directions and fit either a single-variate
         variogram or an LMC structure (depending on the number of id's).
         variable \code{dir} is available for the purpose of fine tunning, i.e. different variogram for
         a different direction.
         In the backend, this code relies on the \code{fit.variogram} routine from the \code{gstat} package.
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}
