% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_functions.R
\name{fstat}
\alias{fstat}
\title{Creates an \code{fstat} structure.}
\usage{
fstat(g = NULL, vName, Coordinates, Functions, scalar = FALSE)
}
\arguments{
\item{g}{- An existing fstat structure (optional). Defaults to \code{NULL}.}

\item{vName}{- Variable name. Note this will be the name used in all analyses.}

\item{Coordinates}{- All coordinates/covariates associated with the functional data (a \code{data.frame}).}

\item{Functions}{- A data frame with functional data stacked in columns.}

\item{scalar}{- Scalars or Functions? (default: \code{scalar = FALSE}). Note: this will determine forecasting, variograms, etc.}
}
\details{
Returns an \code{fstat} structure which is a simple \code{R} list consisting of the following slots:
a) \code{data} - a list with all data, each element has two slots: coordinates and functions;
b) \code{variograms} - an empty slot reserved for empirical variograms computed with \code{fvariogram}
c) \code{models}     - an empty slot reserved for variogram models computed with \code{fit.variogram} or \code{fit.lmc}

When adding functional data, the code will check whether covariates passed in coordinates data frame match the covariates
in the coordinates data frame that is already present at slot \code{$data[[1]]$coordinates}. If they don't an error will pop up.
Extra covariates are removed without prompt. You've been warned!

Note that if parameter scaling (i.e. unit cube) is required, that should be performed as a part of pre-processing.
see \code{help(scaleInput)} function of this package.
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu}), Alessandra Menafoglio (\email{alessandra.menafoglio@polimi.it})
}
