% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_functions.R
\name{gstatify}
\alias{gstatify}
\title{A function that converts Cpp output to gstat variogram type class:}
\usage{
gstatify(M = NULL, label = NULL, dir.label = NULL)
}
\arguments{
\item{M}{- A matrix produced by one of the Cpp variogram computing routines (i.e. \code{empTraceVariogram});}

\item{label}{- A label you wish to assign as \code{id} in variogram structure (The Cpp code does not return any variable id);}

\item{dir.label}{- A vector of labels for directional variogram. The Cpp code returns [0,1,2,...].}
}
\value{
A data frame of class: \code{gstatVariogram}.
}
\details{
Not to be called directly! Serves as a suport function only!
         We expose it anyways because it might be useful in research.
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}
