% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_functions.R
\name{predictFstat}
\alias{predictFstat}
\title{A general predict function for fstat structure}
\usage{
predictFstat(.g, .newCoordinates = NULL, .what = NULL, .type = "OK",
  ErrorCheck = FALSE, algIndependent = FALSE)
}
\arguments{
\item{.g}{- an fstat structure}

\item{.what}{- variable you wish to predict (must match one of the already present names)}

\item{.type}{- kriging type any of the following: 'OK', 'UK', 'OcoK', and 'UcoK'. Applies to both scalar and functional cases.}

\item{ErrorCheck}{- whether to check for errors on each segment and print info.}

\item{algIndependent}{- Boolean, specifying whether drifts in cokriging are independent (default = FALSE)}
}
\details{
In all methods we use covariances (no variograms) by subtracting the variogram from the sill. The type of covariance structure is
         extracted from the fstat structure.

         We don't provide functionality for simple kriging and simple co-kriging since doing them is just silly.

         UcoK can have three types of drifts. Algebraically independent, algebraically dependent and mixed (see Chiles and Delfiner, 1998)
         This code implements the first two cases (algIndependent flag). Mixed drift will be implemented in one of the future versions of this code.
}
\author{
Ogy Grujic \email{ogyg@stanford.edu}
}
