% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{scaleInput}
\alias{scaleInput}
\title{A simple scaling of a vector of parameters (i.e. to (0,1) scale)}
\usage{
scaleInput(.Input, .lB = NULL, .uB = NULL)
}
\arguments{
\item{.Input}{- Input vector you wish to scale.}

\item{.lB}{- lower bound you wish to use for scaling (default = min(.Input))}

\item{.uB}{- upper bound you wish to use for scaling (default = max(.Input))}
}
\author{
Ogy Grujic
}
