% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary.fstat}
\alias{summary.fstat}
\title{Plots summary statistics for regression model(s)}
\usage{
\method{summary}{fstat}(.g = NULL)
}
\arguments{
\item{.g}{- An fstat structure}
}
\author{
Ogy Grujic (ogyg@stanford.edu)
}
