% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_functions.R
\name{traceKrigeCV}
\alias{traceKrigeCV}
\title{Performs k-fold cross validation on the kriging model:}
\usage{
traceKrigeCV(.coordinates, .functions, .model, .F = NULL)
}
\arguments{
\item{Coordinates}{- Coordinates of your functional data}

\item{Functions}{- Functions stacked in columns}

\item{Variogram}{Model - Class vgm, produced as a call to gstat::vgm}

\item{.covType}{- Covariance type: "omni", "product", "sum".}

\item{F}{- Universal kriging matrix. For ordinary kriging a vector of 1's.
For simple kriging \code{NULL} which is default!}
}
\details{
This function performs k-fold cross validation for a given variogram model
         If F=NULL it assumes simple kriging in which case it must work with covariances.
         Covariances are produced from the variogram by substracting the variogram values
         from the fitted sill. Of course, this assumes second order stationarity! It's users
         responsibility to verify that this is truly appropriate!

         In all other cases of universal kriging (OK as a special case), the code will
         work with variograms.

         Note: This is an experimental function! Use at your own responsibility!
}
\author{
Ogy Grujic (\email{ogyg@stanford.edu})
}
