% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trapzc}
\alias{trapzc}
\title{Numerical integration via trapezoid rule}
\usage{
trapzc(step, y)
}
\arguments{
\item{step}{- Step along the abscissa}

\item{y}{- Function evaluated on an equispaced grid with \code{step = step}}
}
\value{
Integral of \code{y}
}
\details{
This is a support function for \code{\link{gammaTraceEmp}}. As such it should not be called directly!
}
\references{
\url{https://en.wikipedia.org/wiki/Trapezoidal_rule}
}
\author{
Alessandra Menafoglio (\email{alessandra.menafoglio@polimi.it})
}
